Starting playing with sqlalchemy
